% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityreg-package.R
\docType{package}
\name{mobilityreg-package}
\alias{mobilityreg}
\alias{mobilityreg-package}
\title{mobilityreg: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
